Fix type cast warnings

+++ lock.h
@@ -18,7 +18,7 @@ static inline uint64_t l_lock_(Lock* l, void* ptr, uint64_t version) {
   while (1) {
     v = atomic_read(&l->value);
     if ((v & 1) ? (v > version) : (v != (uint64_t)ptr)) return 0;
-    if (atomic_cmpswp(&l->value, v, ptr)) return v;
+    if (atomic_cmpswp(&l->value, v, (uint64_t)ptr)) return v;
   }
 }
 
@@ -34,7 +34,7 @@ static inline int l_lock  (Lock* l, void* ptr, uint64_t version) {
   while (1) {
     v = atomic_read(&l->value);
     if ((v & 1) ? (v > version) : (v != (uint64_t)ptr)) return 0;
-    if (atomic_cmpswp(&l->value, v, ptr)) return 1 + (v == (uint64_t)ptr);
+    if (atomic_cmpswp(&l->value, v, (uint64_t)ptr)) return 1 + (v == (uint64_t)ptr);
   }
 }
